improve dom structure
