+++ data/utils.ts
@@ -1,19 +1,11 @@
-import { fromBinary } from '@bufbuild/protobuf'
-import {
-  ModuleInfoSchema,
-  ModuleInfo as StardocModuleInfo,
-} from '@buf/bazel_bazel.bufbuild_es/src/main/java/com/google/devtools/build/skydoc/rendering/proto/stardoc_output_pb.js'
 import path from 'path'
 import { formatISO, parse } from 'date-fns'
 import { execa } from 'execa'
-import { promises as fs } from 'node:fs'
+import { promises as fs } from 'fs'
 import { gitlogPromise } from 'gitlog'
 import * as os from 'os'
 import pMemoize from 'p-memoize'
 import * as yaml from 'js-yaml'
-import * as tar from 'tar-stream'
-import * as zlib from 'zlib'
-import { Readable } from 'node:stream'
 
 export const MODULES_ROOT_DIR = path.join(
   process.cwd(),
@@ -278,7 +270,6 @@ const buildAllModuleInfoInner = async (): Promise<AllModuleInfo> => {
         allModuleInfo.reverseDependencies[dependency.module] ||= new Set()
         allModuleInfo.reverseDependencies[dependency.module].add(moduleName)
       }
-      break
     }
   }
 
@@ -493,108 +484,5 @@ export const getSourceJson = async (
     return JSON.parse(sourceContents) as SourceJson
   } catch {
     return null
-  }
-}
-
-export const fetchDocsList = async (docsUrl: string): Promise<string[]> => {
-  if (typeof window !== 'undefined') {
-    throw new Error('fetchDocsList should only be called server-side')
-  }
-
-  try {
-    const response = await fetch(docsUrl)
-    if (!response.ok) {
-      console.warn('Failed to fetch docs from ', docsUrl, response)
-      return []
-    }
-
-    const docsArchive = await response.arrayBuffer()
-    console.warn('222 docsArchive', docsArchive)
-    return new Promise((resolve, reject) => {
-      const extract = tar.extract()
-      const stardocs: string[] = []
-      console.warn('111 extract')
-
-      extract.on('entry', (header, stream, next) => {
-        console.warn('333 stream - processing entry:', header.name, header.type)
-        const chunks: any[] = []
-
-        stream.on('data', (chunk) => {
-          console.warn(
-            '444 data chunk received for:',
-            header.name,
-            chunk.length
-          )
-          chunks.push(chunk)
-        })
-
-        stream.on('end', () => {
-          try {
-            const content = Buffer.concat(chunks)
-            console.warn(
-              '555 end - processing file:',
-              header.name,
-              'size:',
-              content.length
-            )
-
-            if (
-              header.type === 'file' &&
-              header.name.endsWith('.binaryproto')
-            ) {
-              const stardoc = fromBinary(
-                ModuleInfoSchema,
-                new Uint8Array(content)
-              )
-              stardocs.push(stardoc.moduleDocstring)
-              console.warn(
-                '666 pushed stardoc:',
-                header.name,
-                stardoc.moduleDocstring
-              )
-            }
-            next()
-          } catch (err) {
-            console.warn('777 error processing file:', header.name, err)
-            next() // Continue processing other files even if one fails
-          }
-        })
-
-        stream.on('error', (err) => {
-          console.warn('888 stream error for:', header.name, err)
-          next() // Continue processing other files
-        })
-      })
-
-      extract.on('finish', () => {
-        console.warn(
-          '999 extraction finished, found',
-          stardocs.length,
-          'stardocs'
-        )
-        resolve(stardocs)
-      })
-
-      extract.on('error', (err) => {
-        console.warn('000 extract error:', err)
-        reject(err)
-      })
-
-      // Turn the Uint8Array into a Readable stream and pipe through gunzip + tar
-      console.warn(
-        '111 starting pipe from docsArchive, size:',
-        docsArchive.byteLength
-      )
-      Readable.from([Buffer.from(docsArchive)])
-        .pipe(zlib.createGunzip())
-        .pipe(extract)
-        .on('error', (err) => {
-          console.warn('222 pipe error:', err)
-          reject(err)
-        })
-    })
-  } catch (error) {
-    console.warn('333 error', error)
-    return []
   }
 }
