+++ package.json
@@ -44,6 +44,7 @@
     "eslint-config-next": "15.5.3",
     "postcss": "~8.4.14",
     "prettier": "^3.0.0",
+    "react-markdown": "^10.1.0",
     "tailwindcss": "~3.1.3",
     "typescript": "5.2.2",
     "zlib": "^1.0.5"
